+++ src/request.rs
@@ -174,6 +174,12 @@ impl HttpRequest {
         self.url_for(name, &NO_PARAMS)
     }
 
+    #[inline]
+    /// Get a reference to a `ResourceMap` of current application.
+    pub fn resource_map(&self) -> &ResourceMap {
+        &self.0.rmap
+    }
+
     /// Peer socket address
     ///
     /// Peer address is actual socket address, if proxy is used in front of