+++ CHANGES.md
@@ -6,9 +6,13 @@
 
 * Add `Responder` impl for `(T, StatusCode) where T: Responder`
 
+* Allow to access app's resource map via
+  `ServiceRequest::resource_map()` and `HttpRequest::resource_map()` methods.
+
 ### Changed
 
 * Upgrade `rand` dependency version to 0.7
+
 
 ## [1.0.3] - 2019-06-28